--- .vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "python.pythonPath": "venv/bin/python3.6"
-}
