+++ apocalypse_sim/run.py
@@ -8,14 +8,12 @@ import os
 import sys
 import getopt
 
-help_text = "Choose your mode with --,\n\
-            --test: runs test maps,\n\
-            --netherland: run the holland map\
-            --experiments: experiment field\
-            --doorway: doorway tests\
-            --roads: map where zombies can't take roads\
-            --situation: all flocking components tested\
-            "
+help_text = ("\nChoose your mode with --:\n" +
+             "--test: runs visual inspection maps\n" +
+             "--netherland: run the holland map\n" +
+             "--doorway: doorway tests\n" +
+             "--experiments: map with parameters that match parameters in " +
+             "our experiments")
 
 if sys.argv[-1] == "--test":
     os.environ["mode"] = "1"
@@ -25,10 +23,8 @@ elif sys.argv[-1] == "--experiments":
     os.environ["mode"] = "3"
 elif sys.argv[-1] == "--doorway":
     os.environ["mode"] = "4"
-elif sys.argv[-1] == "--roads":
-    os.environ["mode"] = "5"
 elif sys.argv[-1] == "--situation":
-    os.environ["mode"] = "6"
+    os.environ["mode"] = "5"
 elif sys.argv[-1] == "--help":
     print(help_text)
     exit(0)
